+++ src/app/admin/(configuration)/parameters/types/index.ts
@@ -1,18 +1,18 @@
 /// CRUD de parametros
 
 export interface ParametroCRUDType {
-  id: string
-  code: string
-  name: string
-  group: string
-  description: string
-  status: string
+  id: string;
+  code: string;
+  name: string;
+  group: string;
+  description: string;
+  status: string;
 }
 
 export interface CrearEditarParametroCRUDType {
-  id?: string
-  code: string
-  name: string
-  group: string
-  description: string
+  id?: string;
+  code: string;
+  name: string;
+  group: string;
+  description: string;
 }
